chore: Expose test stuff via index.js
Properly expose test helper exports
Change how we mock kint components
