+++ test/jest/helpers/index.js
@@ -2,3 +2,4 @@ export { default as TestForm } from './TestForm';
 export { default as Harness } from './Harness';
 export { default as renderWithIntl } from './renderWithIntl';
 export { default as mockErmComponents } from './mockErmComponents';
+export { default as mockKintComponents } from './mockKintComponents';
