+++ test/jest/helpers/mockKintComponents.js
@@ -1,0 +1,15 @@
+const mockKintComponents = {
+  useRefdata: jest.fn().mockReturnValue([]),
+  useCustomProperties: jest.fn(({
+    returnQueryObject = false
+  }) => {
+    let returnShape = [];
+    if (returnQueryObject) {
+      returnShape = {};
+    }
+
+    return ({ data: returnShape, isLoading: false });
+  }),
+};
+
+export default mockKintComponents;
